# [pickers] Improve performances of `DayCalendar`