--- a/docs/data/date-pickers/date-picker/CustomDay.js
+++ b/docs/data/date-pickers/date-picker/CustomDay.js
@@ -33,14 +33,14 @@ const CustomPickersDay = styled(PickersDay, {
 }));
 
 function Day(props) {
-  const { day, selectedDays } = props;
+  const { day, selectedDay } = props;
 
-  if (selectedDays.length === 0) {
+  if (selectedDay == null) {
     return <PickersDay {...props} />;
   }
 
-  const start = selectedDays[0].startOf('week');
-  const end = selectedDays[0].endOf('week');
+  const start = selectedDay.startOf('week');
+  const end = selectedDay.endOf('week');
 
   const dayIsBetween = day.isBetween(start, end, null, '[]');
   const isFirstDay = day.isSame(start, 'day');
@@ -62,19 +62,24 @@ Day.propTypes = {
    * The date to show.
    */
   day: PropTypes.object.isRequired,
-  /**
-   * Currently selected days.
-   */
-  selectedDays: PropTypes.arrayOf(PropTypes.object).isRequired,
+  selectedDay: PropTypes.object,
 };
 
 export default function CustomDay() {
+  const [value, setValue] = React.useState(dayjs('2022-04-07'));
+
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
       <StaticNextDatePicker
         displayStaticWrapperAs="desktop"
-        defaultValue={dayjs('2022-04-07')}
+        value={value}
+        onChange={(newValue) => setValue(newValue)}
         components={{ Day }}
+        componentsProps={{
+          day: {
+            selectedDay: value,
+          },
+        }}
       />
     </LocalizationProvider>
   );