--- a/docs/data/date-pickers/date-picker/CustomDay.tsx
+++ b/docs/data/date-pickers/date-picker/CustomDay.tsx
@@ -37,15 +37,15 @@ const CustomPickersDay = styled(PickersDay, {
   }),
 })) as React.ComponentType<CustomPickerDayProps>;
 
-function Day(props: PickersDayProps<Dayjs>) {
-  const { day, selectedDays } = props;
+function Day(props: PickersDayProps<Dayjs> & { selectedDay?: Dayjs | null }) {
+  const { day, selectedDay } = props;
 
-  if (selectedDays.length === 0) {
+  if (selectedDay == null) {
     return <PickersDay {...props} />;
   }
 
-  const start = selectedDays[0].startOf('week');
-  const end = selectedDays[0].endOf('week');
+  const start = selectedDay.startOf('week');
+  const end = selectedDay.endOf('week');
 
   const dayIsBetween = day.isBetween(start, end, null, '[]');
   const isFirstDay = day.isSame(start, 'day');
@@ -63,12 +63,20 @@ function Day(props: PickersDayProps<Dayjs>) {
 }
 
 export default function CustomDay() {
+  const [value, setValue] = React.useState<Dayjs | null>(dayjs('2022-04-07'));
+
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
       <StaticNextDatePicker
         displayStaticWrapperAs="desktop"
-        defaultValue={dayjs('2022-04-07')}
+        value={value}
+        onChange={(newValue) => setValue(newValue)}
         components={{ Day }}
+        componentsProps={
+          {
+            selectedDay: value,
+          } as any
+        }
       />
     </LocalizationProvider>
   );