--- a/docs/data/date-pickers/date-picker/CustomDay.tsx.preview
+++ b/docs/data/date-pickers/date-picker/CustomDay.tsx.preview
@@ -1,5 +1,11 @@
 <StaticNextDatePicker
   displayStaticWrapperAs="desktop"
-  defaultValue={dayjs('2022-04-07')}
+  value={value}
+  onChange={(newValue) => setValue(newValue)}
   components={{ Day }}
+  componentsProps={{
+    day: {
+      selectedDay: value,
+    } as any,
+  }}
 />