--- a/docs/data/date-pickers/legacy-date-picker/CustomDay.js
+++ b/docs/data/date-pickers/legacy-date-picker/CustomDay.js
@@ -34,14 +34,14 @@ const CustomPickersDay = styled(PickersDay, {
 }));
 
 function Day(props) {
-  const { day, selectedDays } = props;
+  const { day, selectedDay } = props;
 
-  if (selectedDays.length === 0) {
+  if (selectedDay == null) {
     return <PickersDay {...props} />;
   }
 
-  const start = selectedDays[0].startOf('week');
-  const end = selectedDays[0].endOf('week');
+  const start = selectedDay.startOf('week');
+  const end = selectedDay.endOf('week');
 
   const dayIsBetween = day.isBetween(start, end, null, '[]');
   const isFirstDay = day.isSame(start, 'day');
@@ -63,10 +63,7 @@ Day.propTypes = {
    * The date to show.
    */
   day: PropTypes.object.isRequired,
-  /**
-   * Currently selected days.
-   */
-  selectedDays: PropTypes.arrayOf(PropTypes.object).isRequired,
+  selectedDay: PropTypes.object,
 };
 
 export default function CustomDay() {
@@ -81,8 +78,11 @@ export default function CustomDay() {
         onChange={(newValue) => {
           setValue(newValue);
         }}
-        components={{
-          Day,
+        components={{ Day }}
+        componentsProps={{
+          day: {
+            selectedDay: value,
+          },
         }}
         renderInput={(params) => <TextField {...params} />}
         inputFormat="'Week of' MMM d"