--- a/docs/data/date-pickers/legacy-date-picker/CustomDay.tsx
+++ b/docs/data/date-pickers/legacy-date-picker/CustomDay.tsx
@@ -38,15 +38,15 @@ const CustomPickersDay = styled(PickersDay, {
   }),
 })) as React.ComponentType<CustomPickerDayProps>;
 
-function Day(props: PickersDayProps<Dayjs>) {
-  const { day, selectedDays } = props;
+function Day(props: PickersDayProps<Dayjs> & { selectedDay?: Dayjs | null }) {
+  const { day, selectedDay } = props;
 
-  if (selectedDays.length === 0) {
+  if (selectedDay == null) {
     return <PickersDay {...props} />;
   }
 
-  const start = selectedDays[0].startOf('week');
-  const end = selectedDays[0].endOf('week');
+  const start = selectedDay.startOf('week');
+  const end = selectedDay.endOf('week');
 
   const dayIsBetween = day.isBetween(start, end, null, '[]');
   const isFirstDay = day.isSame(start, 'day');
@@ -75,9 +75,12 @@ export default function CustomDay() {
         onChange={(newValue) => {
           setValue(newValue);
         }}
-        components={{
-          Day,
-        }}
+        components={{ Day }}
+        componentsProps={
+          {
+            selectedDay: value,
+          } as any
+        }
         renderInput={(params) => <TextField {...params} />}
         inputFormat="'Week of' MMM d"
       />