--- a/packages/x-date-pickers-pro/src/DateRangeCalendar/dateRangeCalendarClasses.ts
+++ b/packages/x-date-pickers-pro/src/DateRangeCalendar/dateRangeCalendarClasses.ts
@@ -8,6 +8,8 @@ export interface DateRangeCalendarClasses {
   root: string;
   /** Styles applied to the container of a month. */
   monthContainer: string;
+  /** Styles applied to the day calendar container when dragging */
+  dayDragging: string;
 }
 
 export type DateRangeCalendarClassKey = keyof DateRangeCalendarClasses;
@@ -17,5 +19,5 @@ export const getDateRangeCalendarUtilityClass = (slot: string) =>
 
 export const dateRangeCalendarClasses: DateRangeCalendarClasses = generateUtilityClasses(
   'MuiDateRangeCalendar',
-  ['root', 'monthContainer'],
+  ['root', 'monthContainer', 'dayDragging'],
 );