--- a/packages/x-date-pickers-pro/src/DateRangePickerDay/DateRangePickerDay.test.tsx
+++ b/packages/x-date-pickers-pro/src/DateRangePickerDay/DateRangePickerDay.test.tsx
@@ -12,7 +12,6 @@ describe('<DateRangePickerDay />', () => {
   describeConformance(
     <DateRangePickerDay
       day={adapterToUse.date()}
-      selectedDays={[]}
       outsideCurrentMonth={false}
       selected
       onDaySelect={() => {}}