--- a/packages/x-date-pickers/src/DateCalendar/DateCalendar.tsx
+++ b/packages/x-date-pickers/src/DateCalendar/DateCalendar.tsx
@@ -462,6 +462,8 @@ export const DateCalendar = React.forwardRef(function DateCalendar<TDate>(
     prevOpenViewRef.current = openView;
   }, [focusedView, handleFocusedViewChange, openView]);
 
+  const selectedDays = React.useMemo(() => [value], [value]);
+
   return (
     <DateCalendarRoot
       ref={ref}
@@ -528,7 +530,7 @@ export const DateCalendar = React.forwardRef(function DateCalendar<TDate>(
               onMonthSwitchingAnimationEnd={onMonthSwitchingAnimationEnd}
               onFocusedDayChange={changeFocusedDay}
               reduceAnimations={reduceAnimations}
-              selectedDays={[value]}
+              selectedDays={selectedDays}
               onSelectedDaysChange={handleSelectedDayChange}
               shouldDisableDate={shouldDisableDate}
               shouldDisableMonth={shouldDisableMonth}