… docs/data/migration/migration-pickers-v5/migration-pickers-v5.md
… docs/pages/x/api/date-pickers/date-range-calendar.json
… docs/pages/x/api/date-pickers/date-range-picker-day.json
… docs/pages/x/api/date-pickers/pickers-day.json
… docs/translations/api-docs/date-pickers/date-range-calendar-pt.json
… docs/translations/api-docs/date-pickers/date-range-calendar-zh.json
… docs/translations/api-docs/date-pickers/date-range-calendar.json
… docs/translations/api-docs/date-pickers/date-range-picker-day-pt.json
… docs/translations/api-docs/date-pickers/date-range-picker-day-zh.json
… docs/translations/api-docs/date-pickers/date-range-picker-day.json
… docs/translations/api-docs/date-pickers/pickers-day-pt.json
… docs/translations/api-docs/date-pickers/pickers-day-zh.json
… docs/translations/api-docs/date-pickers/pickers-day.json
… packages/x-date-pickers-pro/src/DateRangeCalendar/DateRangeCalendar.test.tsx
… packages/x-date-pickers-pro/src/DateRangePickerDay/DateRangePickerDay.tsx
… packages/x-date-pickers-pro/src/DateRangePickerDay/dateRangePickerDayClasses.ts
… packages/x-date-pickers/src/DateCalendar/DayCalendar.tsx
… packages/x-date-pickers/src/DateCalendar/tests/DateCalendar.test.tsx
… packages/x-date-pickers/src/DateCalendar/useCalendarState.tsx
… packages/x-date-pickers/src/PickersDay/PickersDay.spec.tsx
… packages/x-date-pickers/src/PickersDay/PickersDay.test.tsx
… packages/x-date-pickers/src/PickersDay/PickersDay.tsx
… packages/x-date-pickers/src/internals/index.ts