Fixes #7065

- [x] No prop of a given day should change across renders on `DateCalendar` when we select another day
- [x] No prop of a given day should change across renders on `DateCalendar` when we select another day (a few exceptions are explained below)
- [x] Add tests
- [x] Benchmark more precisely the gains

These performance improvements are a trade off, to achieve them, I had to remove the `selectedDays` prop that was by definition causing a re-render of all the days whenever the value change. I updated the doc demo to show how users can do without it.

## Benchmarks

All the benchmarks below have been realized without any custom props `<DateCalendar />` / `<DateRangeCalendar />`, on my desktop computer with a 6x CPU throttling.

### Date Calendar - click on a date (no prev value)

Before: 347m (2 render: 183ms + 164ms)
After: 128ms (2 render: 67ms + 61ms)

### Date Range Calendar - hover a date

Before: 494ms
After: 105ms 

Do we really need this render ? We have none on the `DateCalendar` and their is not preview to render anyway, we should be able to skip it.

### Date Range Calendar - click on a start date (no prev value)

Before: 904ms (2 render: 460ms + 444ms)
After: 239ms (2 render: 130ms + 109ms)

### Date Range Calendar - start dragging

Before: 429ms
After: 121ms

### Date Range Calendar - move dragging from a date to the next one

Before: 837ms (409ms + 418ms)
After: 104ms