@flaviendelangle 2022-12-02
> Even before the last commit the date range selection and dragging had near flawless performance on my mobile.

The last commit is mostly desktop oriented
I did some desktop tests with x6 throttling, and the render when start dragging went from 278ms to about 60ms, it's worth testing more but I think the additional complexity is worth it.
For the bugs, I'll dive deeper, I have a few failing tests and I need to find where I made mistakes.